fix: S12.26 keep (int32_t) cast inside _Report wrapper for clang
clang -Wsign-conversion treats the unscoped error enums' underlying type as
unsigned, so the enum->int32_t conversion must stay explicit. Move the cast
from the (removed) call sites into the one wrapper body per class — call sites
stay cast-free, 2.4 still fixed by the typed param. clang/gcc/cppcheck green.

+++ Core/Source/SolidSyslogBlockStorePrivate.h
@@ -37,7 +37,7 @@ static inline void BlockStore_Report(
     enum SolidSyslogBlockStoreErrors code
 )
 {
-    SolidSyslog_Error(severity, &BlockStoreErrorSource, category, code);
+    SolidSyslog_Error(severity, &BlockStoreErrorSource, category, (int32_t) code);
 }
 
 #endif /* SOLIDSYSLOGBLOCKSTOREPRIVATE_H */
