+++ Core/Source/SolidSyslogPassthroughBufferPrivate.h
@@ -25,7 +25,7 @@ static inline void PassthroughBuffer_Report(
     enum SolidSyslogPassthroughBufferErrors code
 )
 {
-    SolidSyslog_Error(severity, &PassthroughBufferErrorSource, category, code);
+    SolidSyslog_Error(severity, &PassthroughBufferErrorSource, category, (int32_t) code);
 }
 
 #endif /* SOLIDSYSLOGPASSTHROUGHBUFFERPRIVATE_H */
