+++ Core/Source/SolidSyslogStreamSenderPrivate.h
@@ -27,7 +27,7 @@ static inline void StreamSender_Report(
     enum SolidSyslogStreamSenderErrors code
 )
 {
-    SolidSyslog_Error(severity, &StreamSenderErrorSource, category, code);
+    SolidSyslog_Error(severity, &StreamSenderErrorSource, category, (int32_t) code);
 }
 
 #endif /* SOLIDSYSLOGSTREAMSENDERPRIVATE_H */
