+++ Platform/Atomics/Source/SolidSyslogStdAtomicCounterPrivate.h
@@ -24,7 +24,7 @@ static inline void StdAtomicCounter_Report(
     enum SolidSyslogStdAtomicCounterErrors code
 )
 {
-    SolidSyslog_Error(severity, &StdAtomicCounterErrorSource, category, code);
+    SolidSyslog_Error(severity, &StdAtomicCounterErrorSource, category, (int32_t) code);
 }
 
 #endif /* SOLIDSYSLOGSTDATOMICCOUNTERPRIVATE_H */
