+++ Platform/LwipRaw/Source/SolidSyslogLwipRawTcpStreamPrivate.h
@@ -41,7 +41,7 @@ static inline void LwipRawTcpStream_Report(
     enum SolidSyslogLwipRawTcpStreamErrors code
 )
 {
-    SolidSyslog_Error(severity, &LwipRawTcpStreamErrorSource, category, code);
+    SolidSyslog_Error(severity, &LwipRawTcpStreamErrorSource, category, (int32_t) code);
 }
 
 #endif /* SOLIDSYSLOGLWIPRAWTCPSTREAMPRIVATE_H */
