+++ Platform/Posix/Source/SolidSyslogPosixFilePrivate.h
@@ -23,7 +23,7 @@ static inline void PosixFile_Report(
     enum SolidSyslogPosixFileErrors code
 )
 {
-    SolidSyslog_Error(severity, &PosixFileErrorSource, category, code);
+    SolidSyslog_Error(severity, &PosixFileErrorSource, category, (int32_t) code);
 }
 
 #endif /* SOLIDSYSLOGPOSIXFILEPRIVATE_H */
